Inttypes rework (#36)
* Signal int types changes

* vardata_t improvements

* int types consistency rework

* Minor corrections

* Introduction of id types

* replace AVR_ID

Replacing AVR_ID/CTL_id macro function with chr_to_id function

* more pin_id_t additions

* ctlreq rework

* delays and frequencies types rework

* More delays and frequencies type rework

* Voltage ref indexing rework

* Misc int types rework

* removing log files

* sim_id_t renamed

id_t renamed to sim_id_t due to conflict with stdlib on unix

* misc changes

+++ bindings/src/utils/buffer_utils.h
@@ -36,9 +36,9 @@ extern "C" {
  * into an C array.
  * buf is a pointer to a array pointer, that is allocated/freed on demand
  */
-uint32_t import_from_pybuffer(const sipAPIDef* sipAPI,
-                              uint8_t **data,
-                              PyObject* exporter);
+Py_ssize_t import_from_pybuffer(const sipAPIDef* sipAPI,
+                                uint8_t **data,
+                                PyObject* exporter);
 
 /*
  * Helper that exports a block of bytes in memory to a Python buffer exporter
@@ -47,7 +47,7 @@ uint32_t import_from_pybuffer(const sipAPIDef* sipAPI,
  */
 PyObject* export_to_pybuffer(const sipAPIDef* sipAPI,
                              uint8_t *data,
-                             uint32_t len);
+                             Py_ssize_t len);
 
 /*
  * Helper function for importing data from a fixed-length
